commit by himanshu on 30/01/2023

+++ src/main/java/com/emp/repository/EmployeeRepository.java
@@ -1,11 +1,15 @@
 package com.emp.repository;
 
-import com.emp.entity.Employee;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+import com.emp.entity.Employee;
+
+public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
 
 	Employee findByEmail(String email);
 
 	Employee findByEmailOrNumber(String email, String number);
+	
+	
 }
